pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sw-crew"
version = "0.1.0"
description = "Autonomous multi-agent software development crew"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.39.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "click>=8.1.0",
    "gitpython>=3.1.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8.0",
]
phase2 = [
    "chromadb>=0.5.0",
]

[project.scripts]
crew = "crew.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/crew"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
